requirements-freeze.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-swagger==1.0.16
aiosignal==1.3.1
anyio==3.7.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
dill==0.3.6
distro==1.8.0
entrypoints==0.4
fastapi==0.98.0
filelock==3.12.2
flake8==6.0.0
flatbuffers==23.5.26
flatdict==4.0.1
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.7.0
isort==5.12.0
iterative-telemetry==0.0.8
jax==0.4.13
Jinja2==3.1.2
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.2.0
mlem==0.4.13
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pip==22.3.1
platformdirs==3.8.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydantic==1.10.9
pyflakes==3.0.1
Pygments==2.15.1
pyparsing==3.1.0
python-dateutil==2.8.2
python-gitlab==3.15.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.4.2
rsa==4.9
scipy==1.11.1
setuptools==65.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
typer==0.9.0
typing_extensions==4.7.0
urllib3==1.26.16
uvicorn==0.22.0
Werkzeug==2.3.6
wheel==0.40.0
wrapt==1.14.1
yarl==1.9.2
zipp==3.15.0
